fix:We fix the components by creatng each comnponent file for each component and setting up the folder
