+++ index.html
@@ -8,6 +8,6 @@
 </head>
 <body>
     <div id="root"></div>
-    <script type="module" src="App.js"></script>
+    <script type="module" src="./src/App.js"></script>
 </body>
 </html>
