+++ src/utils/constant.js
@@ -1,0 +1,4 @@
+
+export const LOGO_URL = "https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fimg.freepik.com%2Fpremium-vector%2Ffood-logo-design_139869-254.jpg%3Fw%3D2000&f=1&nofb=1&ipt=b4f29c3f45279176429808ef2035ca3672e8945903ceec56b2f36531a2f8a749&ipo=images";
+
+export const CDN_URL = "https://media-assets.swiggy.com/swiggy/image/upload/fl_lossy,f_auto,q_auto,w_660/";
