+++ src/utils/mockData.js
@@ -1237,3 +1237,4 @@ export const resList=[
       }
 ]
 
+export default resList;
